Feat: Se agrega pantalla de carga de datos y muestreo para el módulo de carga de empleados o registro de personal

+++ Seph.Principal.Application/Features/HistorialContratos/DTOs/CreateHistorialContratoRequest.cs
@@ -12,7 +12,7 @@ namespace Seph.Principal.Application.Features.HistorialContratos.DTOs
         long IdEmpleado,
         long IdInstitucion,
         long IdTipoPersonal,
-        int IdTipoContrato,
+        long IdTipoContrato,
         string? StrOtroTipoContrato,
         long IdArea,
         DateTime DateTimeFechaRegistro,
